[Refine code]Move class Padding into the NeonDepthwiseConv.h.
